refactor:move dashboard files into dashboard component folder
